Fix zml's issue, AliasContext javadocs, refactor a little.
