+++ src/main/java/org/spongepowered/api/service/command/CommandMessageFormatting.java
@@ -22,12 +22,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.util.command.dispatcher;
+package org.spongepowered.api.service.command;
 
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
 
 class CommandMessageFormatting {
+
     public static Text error(Text error) {
         return error.builder().color(TextColors.RED).build();
     }
