+++ src/main/java/org/spongepowered/api/service/command/TranslationPlaceholder.java
@@ -22,18 +22,52 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.util.command.dispatcher;
+package org.spongepowered.api.service.command;
 
 import org.spongepowered.api.text.Text;
-import org.spongepowered.api.text.format.TextColors;
+import org.spongepowered.api.text.Texts;
+import org.spongepowered.api.text.translation.Translation;
 
-class CommandMessageFormatting {
-    public static Text error(Text error) {
-        return error.builder().color(TextColors.RED).build();
+/**
+ * This class contains methods intended for use as placeholders until true
+ * localization is implemented for SpongeAPI strings.
+ */
+class TranslationPlaceholder {
+
+    private TranslationPlaceholder() {
     }
 
-    public static Text debug(Text debug) {
-        return debug.builder().color(TextColors.GRAY).build();
+    public static Text _(String untranslated, Object... args) {
+        return Texts.of(new FixedTranslation(untranslated), args);
+    }
+
+    private static class FixedTranslation implements Translation {
+
+        private final String untranslated;
+
+        private FixedTranslation(String untranslated) {
+            this.untranslated = untranslated;
+        }
+
+        @Override
+        public String getId() {
+            return this.untranslated;
+        }
+
+        @Override
+        public String getName() {
+            return this.untranslated;
+        }
+
+        @Override
+        public String get() {
+            return this.untranslated;
+        }
+
+        @Override
+        public String get(Object... args) {
+            return String.format(this.untranslated, args);
+        }
     }
 
 }
